+++ packages/css/src/components/icon/icon.template.ts
@@ -2,9 +2,7 @@ import { Icon } from '@dso-toolkit/sources';
 import { html } from 'lit-html';
 
 export function iconTemplate({ icon }: Icon) {
-  return html`
-    <svg class="di di-${icon}">
-      <use href="dso-icons.svg#${icon}" />
-    </svg>
-  `;
+  return html`<svg class="di di-${icon}">
+    <use href="dso-icons.svg#${icon}" />
+  </svg>`;
 }