+++ packages/sources/src/components/viewer-grid/viewer-grid.models.ts
@@ -1,3 +1,16 @@
+import { DefinitionList } from '../..';
+
+export interface ViewerGridDemoProperties {
+  panelSize: 'small' | 'medium' | 'large';
+  panelOpen: boolean;
+  shrink: (e: MouseEvent) => void;
+  expand: (e: MouseEvent) => void;
+  closeOverlay: (e: MouseEvent) => void;
+  documentHeaderFeaturesOpen: boolean;
+  documentHeaderMapAction: (e: MouseEvent) => void;
+  documentHeaderFeatureAction: (e: MouseEvent) => void;
+}
+
 export interface ViewerGrid<TemplateFnReturnType> {
   main: TemplateFnReturnType;
   map: TemplateFnReturnType;
@@ -9,10 +22,20 @@ export interface ViewerGrid<TemplateFnReturnType> {
   closeOverlay: (e: MouseEvent) => void;
 }
 
-export interface ViewerGridDemoProperties {
-  panelSize: 'small' | 'medium' | 'large';
-  panelOpen: boolean;
-  shrink: (e: MouseEvent) => void;
-  expand: (e: MouseEvent) => void;
-  closeOverlay: (e: MouseEvent) => void;
+export interface ViewerGridDocumentHeader<TemplateFnReturnType> {
+  title: TemplateFnReturnType;
+  type: string;
+  owner: string;
+  status: TemplateFnReturnType;
+  features: DefinitionList<TemplateFnReturnType>;
+  featuresOpen: boolean;
+  mapAction: (e: MouseEvent) => void;
+  featuresAction: (e: MouseEvent) => void;
+}
+
+export interface ViewerGridDocumentListItem<TemplateFnReturnType> {
+  title: TemplateFnReturnType;
+  type: string;
+  owner: string;
+  status: TemplateFnReturnType;
 }