+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## NEXT
 
+### Added
+* **css:** document-header en document-list-item in dso-viewer-grid main panel #1308 ([#1308](https://github.com/dso-toolkit/dso-toolkit/issues/1308))
+
 ### Fixed
 * **dso-toolkit + sources:** Placeholder tekst buiten inputs in Safari ([#1364](https://github.com/dso-toolkit/dso-toolkit/issues/1364))
 * **dso-toolkit:** Fractal dev omgeving watcht files uit packages/css niet ([#1370](https://github.com/dso-toolkit/dso-toolkit/issues/1370))